Add Evergreen workflow
I'm very much looking forward to taking this for a spin. Going to run it
with `DRY_RUN` set to `true` so that I can see which repositories will
be affected, then swap out the `ORGANIZATION` environment variable for
`REPOSITORY` and run it on one of our relatively inactive repositories
to see it in action.